requirements.txt
google-generativeai 
supabase 
python-dotenv
fastapi
uvicorn
pydantic